Started on SpellCastingComponent
One of the key features of this game is the spell casting. At this time, the player has a number of mana, access to a deck of spell cards, their hand of spell cards (taken from their deck), and their discard pile. When the player moves, the first spell in their hand is discarded, other hand spells are shifted once to the left, and a new spell is drawn from the deck (plus a number of mana is regenerated if they are below max). Also if they have enough mana the player is able to cast on of their spells. Whenever the player does use one of their spells, its mana cost is deducted from the player's current mana, the spell card is discarded and a new spell from the deck takes its place in the hand. This causes a constant shifting changing of the options the player has available at any given moment, hopefully making them strategize accordingly.