Added a lot: FieldManager and Generator, Basic enemy WIP, deck initialisation
Long overdue commit, I have created functionality to enable/disable the players deck and its capabilities while not in combat, started on implementing a new way to spawn the map and its contents as I am currently now restricting it to a single room that changes, instead of multiple rooms. I have also started working on enemies who currently will move towards the player, one tile at a time, after the player moves. There is currently a bug with this however as when an enemy is blocked by another enemy, in some cases the blocked enemy stops attempting to move. I think those are the main things but there are other smaller tweaks as well.